requirements.txt
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
gunicorn==19.9.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
Pygments==2.4.2
requests==2.22.0
traitlets==4.3.2
wcwidth==0.1.7
Django==3.0.3
django-modeltranslation==0.14.4
django-storages==1.9.1
future==0.18.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.8
pefile==2019.4.18
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7